_controlQueue is now a std::deque
Based on commit f5e439c8f86609c3be855b76844352322e366b80 of dumbfixes
branch of 0xFEEDC0DE64 fork of ESPAsyncWebServer.

Unlike the original patch, clearing the std::queue using assignment of
{} is not supported in C++11, and no other method of clearing the queue
is available, so this forces use of std::deque instead. This in turn
forces all methods of std::queue to be replaced with their underlying
equivalents in std::deque.
Control messages are constructed directly inside the container using
.emplace_back() (.emplace() in the original patch) instead of building
the message and passing its pointer as a parameter.