Add locking around _messageQueue and _controlQueue in AsyncWebSocketClient
Based on a combination of commits 344d31c1ffbffd845823e1f8c369fcbf2439e47d
and 9b98550f64000b937db9567b69e1275feccf1f0f of of dumbfixes branch of
0xFEEDC0DE64 fork of ESPAsyncWebServer.

This patch additionally introduces a PolymorphMessageContainer class,
containing either a AsyncWebSocketBasicMessage or an
AsyncWebSocketMultiMessage, and intented to be contained inside a
std::deque (instead of a std::queue as on the original commits, as
done in a previous commit). This patch therefore eliminates one more use
of a LinkedList<T>.

Some code formatting done on the side too.