Fix self-deadlock onConnect introduced by ul-gh/fix_#884 branch
Commit 344062ae5be9dab80dd602458857f3a5ec5dfb3c intended to protect the
client list for AsyncEventSource from concurrent access by wrapping
accesses with locking. However, this introduces a self-deadlock scenario
if application code in the onConnect callback (now called with the
client list lock held) tries to broadcast a message to existing
connections in the same AsyncEventSource (for example, to alert of a
new incoming connection). The broadcast call tries to take the lock but
blocks because the lock is already taken, by the same task, before
calling the callback. Fixed by making use of the AsyncWebLockGuard class
which is tailor-made to address this scenario.

(cherry picked from commit 646eab7bd8761a7ff89c627e37c0f27c49cef32f)